chore: separate LikeButton into Activity and Plan LikeButton
