+++ src/components/LikeButton/PlanLikeButton/LikeButton.module.css
@@ -1,0 +1,14 @@
+.container {
+    width: 100%;
+    height: 100%;
+    padding: 4px 3px 3px;
+    background-color: white;
+    border: 1 solid black;
+    border-radius: 50%;
+}
+
+.img {
+    width: 100%;
+    height: 100%;
+    cursor: pointer;
+  }
